refactor app packaging
