+++ app/src/main/AndroidManifest.xml
@@ -34,7 +34,7 @@
         tools:replace="android:icon"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".activity.MainActivity"
+            android:name=".ui.activity.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -56,7 +56,7 @@
         </service>
         <!-- ./Firebase Notifications -->
         <activity
-            android:name=".activity.NotificationActivity"
+            android:name=".ui.activity.NotificationActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar"></activity>
     </application>
