+++ app/src/main/java/com/myduka/app/service/MyFirebaseMessagingService.java
@@ -25,9 +25,9 @@ import android.text.TextUtils;
 
 import com.google.firebase.messaging.FirebaseMessagingService;
 import com.google.firebase.messaging.RemoteMessage;
-import com.myduka.app.activity.MainActivity;
-import com.myduka.app.app.Config;
-import com.myduka.app.utils.NotificationUtils;
+import com.myduka.app.ui.activity.MainActivity;
+import com.myduka.app.util.Config;
+import com.myduka.app.util.NotificationUtils;
 
 import org.json.JSONException;
 import org.json.JSONObject;
