+++ app/src/main/java/com/myduka/app/util/Utils.java
@@ -1,4 +1,4 @@
-package com.myduka.app.app;
+package com.myduka.app.util;
 
 import android.util.Base64;
