+++ app/src/main/res/menu/menu_main.xml
@@ -19,7 +19,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.myduka.app.activity.MainActivity">
+    tools:context="com.myduka.app.ui.activity.MainActivity">
     <item
         android:id="@+id/action_checkout"
         android:orderInCategory="100"
